refactor(workflows): centralize types and use MockMcpServer for testing
- Introduced a new types.ts file to define shared interfaces and types for workflows.
- Replaced direct imports of McpServer with MockMcpServer and other types from types.ts.
- Updated all workflow modules to use the centralized types for configs and server.
- Improved error handling to safely stringify errors.
- Enhanced type safety and consistency across workflow modules.
- Updated runWorkflowTests to use MockMcpServer for mocking.

This refactor improves maintainability and testability of the Illustrator workflow codebase.
